Simplify the way the plugin gets extra config
Prior to this I was extending the django-stubs config class, but that
class isn't designed to be extended and it was quite awkward.

It also meant that I had to repeat django-stubs logic in this plugin and
skip django-stubs init.

After this that is no longer necessary and the extra configuration is
retrieved separately to the rest of the django-stubs config.

The downside is that if those extra options are invalid, the plugin will
load django before that is discovered. This may change in the next few
changes