Don't expand additional dependencies
Turns out now that we resolve annotations by using type aliases in the
virtual dependencies that we don't need to provide additional
dependencies for dependencies of the module.

We only need to provide the virtual dependency for that file, and if
we're using an incremental cache, we also make the connection to django
settings explicit.

We had to expand the dependencies in the past because we relied on mypy
time information to resolve the dependencies. But now we only need the
relevant type alias to do that.

And the less we add to additional dependencies, the more performant mypy
becomes. Especially on large projects