Remove `Concrete.type_var`
This changes removes the complexity of `Concrete.type_var` which would
created a TypeVar with each concrete model as a separate choice.

It instead says we should use `TypeVar("T_Model", bound=Concrete[Model])`

This should be enough the same to be fine, is no less useful than
`TypeVar("T_Model", bound=Model)` and has the other nice benefit of
being a real TypeVar that mypy isn't sad about.