Convert from fork of pytest-mypy-plugins to pytest-typing-runner
In the beginning there was pytest-mypy-plugins as used by django-stubs
and I needed the ability to test how mypy incremental and fine-grained
caches interact with my mypy plugin.

I created this issue on pytest-mypy-plugins asking if it were possible
https://github.com/typeddjango/pytest-mypy-plugins/issues/144

The answer was that this functionality wasn't in the plugin and PRs were
welcome

I made a few PRs including
https://github.com/typeddjango/pytest-mypy-plugins/pull/152 which was
then the base for a couple more branches.

Ultimately it was a lot of change and getting that merged wasn't a
thing.

After many months I decided to take what I had done in those branches
and what I had here and create pytest-typing-runner as a more
sophisticated, tested, documented, typed version of what I had written
across those branches and this repo.

The project should be able to handle more than just mypy, but that has
yet to be done. Also I want to make it so that pytest-typing-runner can
be used in django-stubs, which involves a few more additional features.

The docs for the project are also in need of some more effort.

But the pytest-typing-runner plugin is now fully tested and I'm happy
with it's API, so I've released a version and made this project use it